pyproject.toml
[tool.poetry]
name = "chat-room"
version = "0.1.0"
description = "chat room"
authors = ["Tên Bạn <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.3"
psycopg2-binary = "^2.9.9"
djangorestframework = "^3.14.0"
factory-boy = "^3.3.0"
django-cors-headers = "^4.3.1"
django-filter = "^23.2"
pillow = "^10.1.0"
opencv-python = "^4.8.0"
numpy = "^1.24.0"
scikit-learn = "^1.3.0"
djangorestframework-simplejwt = "^5.3.1"
channels = "^4.1.0"
channels-redis = "^4.2.0"
daphne = "^4.1.2"
dj-database-url = "^2.1.0"
twisted = {extras = ["http2", "tls"], version = "^24.3.0"}
boto3 = "^1.34.122"
django-storages = "^1.14.3"
django-environ = "^0.11.2"
python-dotenv = "^1.0.1"

[tool.poetry.dev-dependencies]
pytest = "^7.4.0"

[tool.poetry.group.dev.dependencies]
coverage = "^7.3.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
